% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyVP.R
\name{easyVP}
\alias{easyVP}
\title{easyVP}
\usage{
easyVP(DEGlist, main, FCcol1, FCcol2, FCcol3, FClimit, PVALlimit, LegendPlace)
}
\arguments{
\item{DEGlist}{limma output. List of differentially expressed features with mandatory columns: logFC, P.Value, adj.P.Val, GeneID}

\item{main}{a title for your plot.}

\item{FCcol1}{a first logFC threshold for key color.}

\item{FCcol2}{a second logFC threshold for key color.}

\item{FCcol3}{a third logFC threshold for key color.}

\item{FClimit}{a logFC threshold beyond which dots are annotated with GeneID.}

\item{PVALlimit}{a -log10(P.Value) threshold beyond which dots are annotated with GeneID. Must be -log10(P.Value) and notP.Value or adj.P.Val.}

\item{LegendPlace}{a place to choose for legend between "topleft", "topright", "bottomright", "bottomleft".}
}
\description{
Create annotated volcano plots with colored dots
}
\examples{
MyDEGs <- read.csv2(file="~/MyStudy/MyDEGs.csv", header=T, sep=";",dec=".", row.names=1)
colnames(MyDEGs)[1] <-"logFC"
colnames(MyDEGs)[4] <-"P.Value"
colnames(MyDEGs)[5] <-"adj.P.Val"
colnames(MyDEGs)[8] <-"GeneID"
easyVP(DEGlist=MyDEGs,main="My easy VP", FCcol1=0.5, FCcol2=1, FCcol3=2, FClimit=2, PVALlimit=5, LegendPlace="topright")
}
